backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
apscheduler==3.10.4
firebase-admin==6.5.0
# Core LLM/AI
langchain
openai
google-generativeai
langchain-google-genai
langchain-huggingface
tiktoken

# Environment variables
python-dotenv

# Pydantic for models
pydantic

# HTTP requests
requests

# Google APIs
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Document processing
PyPDF2
pdfplumber
python-docx

# OCR (for image-to-text)
pytesseract
Pillow

# Web scraping (if needed for fetching content)
beautifulsoup4
html2text

# CORS (if your frontend needs to call backend)
fastapi[all]

# Optional: typing extensions for compatibility
typing-extensions
